requirements.txt
apache-airflow
apache-airflow[kubernetes]
SQLAlchemy==1.3.23
